Post: Remove progress bar
Tracking and displaying the reading progress visually with the toggle
dropdown is difficult to do so without compromising aesthetics (as users
might see their progress "jump" when they open/close dropdowns or scroll
from dropdown to dropdown).

An alternative in the future may be to implement smooth animations to
handle these cases, maybe even allow the user to navigate to sections by
clicking on the progress bar.