Home: Implement event card date system (#165)
This event card system uses Ghost's Content API to retrieve a list of posts with the `#event` tag. Afterwards, the excerpt of each post is parsed as a date in ISO format (i.e. YYYY-MM-DD) which is used as the event date.

Currently, the event cards do not link to the event page. This will be changed in the future.